document upload fix

+++ wagtail/admin/views/generic/multiple_upload.py
@@ -142,7 +142,7 @@ class AddView(PermissionCheckedMixin, TemplateView):
             },
             user=request.user,
         )
-        print("HERE")
+
         if form.is_valid():
             # Save it
             self.object = self.save_object(form)
@@ -165,39 +165,43 @@ class AddView(PermissionCheckedMixin, TemplateView):
                 collection_id=self.request.POST.get("collection"),
             )
 
-            # checking for image duplicates, in case custom image model with additional fields is used
             data = self.get_edit_upload_response_data()
-            temporary_image = Image.objects.create(
-                title="TEMPORARY_ORIGINAL_WAGTAIL_IMAGE",
-                file=self.request.FILES["files[]"],
-                width=1,
-                height=1,
-            )
-            temporary_image.get_file_hash()
 
-            duplicates = find_image_duplicates(
-                image=temporary_image,
-                user=self.request.user,
-                permission_policy=self.permission_policy,
-            )
-            temporary_image.delete()
-
-            if not duplicates:
-                data.update(duplicate=False)
-            else:
-                data.update(
-                    duplicate=True,
-                    confirm_duplicate_upload=render_to_string(
-                        "wagtailimages/images/confirm_duplicate_upload.html",
-                        {
-                            "existing_image": duplicates[0],
-                            "delete_action": reverse("wagtailimages:dummy_delete"),
-                        },
-                        request=self.request,
-                    ),
+            if not hasattr(self.request.FILES["files[]"], "image"):
+                return JsonResponse(data)
+            else:  
+                # checking for image duplicates, in case custom image model with additional fields is used
+                temporary_image = Image.objects.create(
+                    title="TEMPORARY_ORIGINAL_WAGTAIL_IMAGE",
+                    file=self.request.FILES["files[]"],
+                    width=1,
+                    height=1,
                 )
+                temporary_image.get_file_hash()
 
-            return JsonResponse(data)
+                duplicates = find_image_duplicates(
+                    image=temporary_image,
+                    user=self.request.user,
+                    permission_policy=self.permission_policy,
+                )
+                temporary_image.delete()
+
+                if not duplicates:
+                    data.update(duplicate=False)
+                else:
+                    data.update(
+                        duplicate=True,
+                        confirm_duplicate_upload=render_to_string(
+                            "wagtailimages/images/confirm_duplicate_upload.html",
+                            {
+                                "existing_image": duplicates[0],
+                                "delete_action": reverse("wagtailimages:dummy_delete"),
+                            },
+                            request=self.request,
+                        ),
+                    )
+
+                return JsonResponse(data)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
